// src/domain/interface/tr-get-me-individual-args.interface.ts
import { TimeIntervalEnum } from 'src/utils/enums/time-interval.enum';
import { TRTimeIntervalEnum } from '../enum/tr-time-interval.enum';
import { User } from 'src/domain/entity/user.entity';

export interface ITRGetMeIndividualArgs {
  startDate?: string;
  endDate?: string;
  timeInterval?: TRTimeIntervalEnum;
  userId?: number;
}

export interface ITRGetMeIndividualValidArgs {
  user: User;
  startDate?: string;
  endDate?: string;
  timeInterval?: TimeIntervalEnum;
}